requirements.txt
alembic==1.10.2
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
bcrypt==4.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.1
decorator==5.1.1
ecdsa==0.18.0
fastapi==0.95.0
flake8==6.0.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
passlib==1.7.4
Paver==1.3.4
pbr==5.11.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pyflakes==3.0.1
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
redis==4.5.4
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.7
SQLAlchemy-serializer==1.4.1
SQLAlchemy-Utils==0.40.0
sqlparse==0.4.3
starlette==0.26.1
Tempita==0.5.2
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.21.1
uvloop==0.17.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==10.4
